+++ src/main/java/com/sparta/deliverybackend/domain/restaurant/controller/RestaurantController.java
@@ -1,6 +1,6 @@
 package com.sparta.deliverybackend.domain.restaurant.controller;
 
-import com.sparta.deliverybackend.api.config.s3.S3Service;
+import com.sparta.deliverybackend.api.s3.service.S3Service;
 import com.sparta.deliverybackend.api.auth.controller.dto.VerifiedMember;
 import com.sparta.deliverybackend.domain.restaurant.controller.dto.*;
 import com.sparta.deliverybackend.domain.restaurant.service.RestaurantService;
